Adds power levels and fixes for INav (#13)
* quick test

* make generic

* remove pitmode from SA when setting 0mW

INav send a 0mW command on boot for some reason and this was forcing the VTx into pitmode.  Removed to prevent pitmode on boot.  mW command is not used by BF and will not affect anything.

* powerLevels

* fix stopbits for INav

* disable pitmode for 0mW command

Do not set pitmode for a 0mW.  INav sends this command on boot for some reason. https://github.com/iNavFlight/inav/issues/6976

* adds inav issue comment

Co-authored-by: Jye <JyeSmith@users.noreply.github.com>